+++ opentelemetry-appender-log/CHANGELOG.md
@@ -4,9 +4,7 @@
 
 ### Fixed
 
-- Populate TimeStamp on LogRecord, instead of ObservedTimeStamp
-  [#1175](https://github.com/open-telemetry/opentelemetry-rust/pull/1175).
-- Add log appender versions to loggers (#1182)
+- Add log appender versions to loggers (#1182).
 
 ## v0.1.0